QuickFit3/lib: project may no longer be saved while it is loaded (bugfix)
QuickFit3/table: 
  - asks now whether to delete a graph
  - changed default font for graphs 
  - bugfix: deleting a graph lead to an endless loop


git-svn-id: https://www.dkfz.de/svn/B040/FCSTOOLS/trunk/QuickFit3@1982 0d8be23c-4016-4ca5-a69c-3788035ecc6c

+++ plugins/table/qfrdrploteditor.cpp
@@ -129,6 +129,7 @@ void QFRDRPlotEditor::addPlot() {
 }
 
 void QFRDRPlotEditor::deleteCurrentPlot() {
+
     deletePlot(tabPlots->currentIndex());
 }
 
@@ -178,10 +179,12 @@ void QFRDRPlotEditor::movePlotDown()
 void QFRDRPlotEditor::deletePlot(int i) {
     QFRDRTable* m=qobject_cast<QFRDRTable*>(current);
     if (m) {
-        updating=true;
-        m->deletePlot(i);
-        rebuildPlotWidgets();
-        updating=false;
+        if (QMessageBox::question(this, tr("table plugin"), tr("Do you really want to delete the current graph?"), QMessageBox::Yes|QMessageBox::No, QMessageBox::No)==QMessageBox::Yes) {
+            updating=true;
+            m->deletePlot(i);
+            rebuildPlotWidgets();
+            updating=false;
+        }
     }
 }
 
@@ -197,9 +200,9 @@ void QFRDRPlotEditor::rebuildPlotWidgets(bool keepPosition) {
     } else {
         while (plotWidgets.size()<current->getPlotCount()) {
             QFRDRTablePlotWidget* w=new QFRDRTablePlotWidget(tabPlots);
-            connect(w, SIGNAL(plotTitleChanged(int,QString)), this, SLOT(updatePlotName(int,QString)));
             plotWidgets.append(w);
             tabPlots->addTab(plotWidgets[plotWidgets.size()-1], "plot");
+            connect(w, SIGNAL(plotTitleChanged(int,QString)), this, SLOT(updatePlotName(int,QString)));
         }
         while (plotWidgets.size()>current->getPlotCount()) {
             int idx=tabPlots->indexOf(plotWidgets.last());
@@ -207,6 +210,7 @@ void QFRDRPlotEditor::rebuildPlotWidgets(bool keepPosition) {
                 tabPlots->removeTab(idx);
             }
             if (plotWidgets.last()) delete plotWidgets.last();
+            plotWidgets.removeLast();
         }
         for (int i=0; i<current->getPlotCount(); i++) {
             //qDebug()<<"   setRecord "<<i;
